% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_complex_conditions}
\alias{df_complex_conditions}
\title{A fictional biological dataset with a complex experimental design}
\format{
a data frame with 360 rows and 4 variables
\itemize{
  \item KO. Boolean value if the sample had a knock out.
  \item DrugA. character vector with "Yes" and "No" elements
    indicating if the sample was treated with drug A.
  \item Timepoint. Numeric vector with elements 8, 24, and 48
    indicating the time of measurement since the beginning of
    the experiment.
  \item response. Numeric vector with the response of the sample
    to the treatment conditions. Could for example be the concentration
    of a metabolite.
}
}
\usage{
df_complex_conditions
}
\description{
A fictional biological dataset with a complex experimental design
}
\examples{
dim(df_complex_conditions)
head(df_complex_conditions)

}
\keyword{datasets}
