% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_pathway_membership}
\alias{gene_pathway_membership}
\title{A fictional dataset describing which genes belong to certain pathways}
\format{
a matrix with 6 rows and 37 columns. Each row is one pathway, with
  its name given as `rownames` and each column is a gene. The values
  in the matrix are Boolean indicators if the gene is a member of the pathway.
}
\usage{
gene_pathway_membership
}
\description{
A fictional dataset describing which genes belong to certain pathways
}
\examples{
dim(gene_pathway_membership)
gene_pathway_membership[, 1:15]
}
\keyword{datasets}
