% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tidy_movies}
\alias{tidy_movies}
\title{Tidy version of the movies dataset from ggplot2 package}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 50000 rows and 10 columns.}
\usage{
tidy_movies
}
\description{
The original \code{\link[ggplot2movies]{movies}} dataset has 7 columns that
contain indicators if a movies belongs to a certain genre. In this version
the 7 columns are collapsed to a single list column to create a tidy
dataset.
}
\keyword{datasets}
